Added json helpers and installation of nlohmann/json.hpp 	modified:   CMakeLists.txt 	modified:   THcParmList.cxx 	renamed:    json.hpp -> nlohmann/json.hpp
